backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.8.0rc2
django==4.2.26
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
idna==3.11
oauthlib==3.3.1
pycparser==2.23
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing-extensions==4.13.2
urllib3==2.2.3
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.8.0rc2
Django==4.2.26
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
idna==3.11
oauthlib==3.3.1
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.2.3
asgiref==3.8.1
babel==2.17.0
backports.zoneinfo==0.2.1
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.8.0rc2
Django==4.2.26
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
docxcompose==1.4.0
docxtpl==0.20.1
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==2.1.5
oauthlib==3.3.1
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.9.0
python-docx==1.1.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.2.3
